requirements.txt
hvac
django
djangorestframework
psycopg2-binary